util: fix macOS build of thread.cc
On macOS builds would fail with:

Undefined symbols for architecture x86_64:
  "kudu::EasyJson& kudu::EasyJson::operator=<unsigned long>(unsigned long)", referenced from:
     kudu::ThreadMgr::ThreadPathHandler(kudu::WebCallbackRegistry::WebRequest const&, kudu::WebCallbackRegistry::WebResponse*) const in thread.cc.o
ld: symbol(s) not found for architecture x86_64
clang: error: linker command failed with exit code 1 (use -v to see invocation)

Rather than instantiating an EasyJson template for unsigned long, which
would be ambiguous, this replaces the usage of size_t with uint64_t.

Reviewed-on: http://gerrit.cloudera.org:8080/14516
Tested-by: Kudu Jenkins

+++ src/kudu/util/thread.cc
@@ -427,7 +427,7 @@ void ThreadMgr::ThreadPathHandler(const WebCallbackRegistry::WebRequest& req,
     }
   } else {
     // List all thread groups and the number of threads running in each.
-    vector<pair<string, size_t>> thread_categories_info;
+    vector<pair<string, uint64_t>> thread_categories_info;
     uint64_t running;
     {
       // See comment above regarding short critical sections.
